Fix possible frequency overflow on Kenwood and Yaesu
This patch fixes a frequency overflow problem on kenwood and yaesu that
can occur with frequencies that over a 32-bit signed int.  This has been
shown to fix the problem on kenwood.

73
Mike W9MDB

+++ kenwood/kenwood.c
@@ -1075,7 +1075,7 @@ int kenwood_set_freq(RIG *rig, vfo_t vfo, freq_t freq)
     rig_debug(RIG_DEBUG_ERR, "%s: unsupported VFO %d\n", __func__, vfo);
     return -RIG_EINVAL;
   }
-  sprintf(freqbuf, "F%c%011ld", vfo_letter, (long)freq);
+  sprintf(freqbuf, "F%c%011lld", vfo_letter, (long long)freq);
 
   int err = kenwood_transaction(rig, freqbuf, NULL, 0);
